+++ utils_test.go
@@ -40,3 +40,209 @@ func TestParseCoord(t *testing.T) {
 	_, err = ParseCoord("P:1:2:3456")
 	assert.NotNil(t, err)
 }
+
+func TestName2id(t *testing.T) {
+	assert.Equal(t, ID(0), name2id("Not valid"))
+	assert.Equal(t, LightFighterID, name2id("Light Fighter"))
+	assert.Equal(t, LightFighterID, name2id("Chasseur léger"))
+	assert.Equal(t, LightFighterID, name2id("Leichter Jäger"))
+	assert.Equal(t, LightFighterID, name2id("Caça Ligeiro"))
+	assert.Equal(t, HeavyFighterID, name2id("Caça Pesado"))
+	assert.Equal(t, LargeCargoID, name2id("Großer Transporter"))
+	assert.Equal(t, DestroyerID, name2id("Zerstörer"))
+	assert.Equal(t, SmallCargoID, name2id("Nave pequeña de carga"))
+	assert.Equal(t, SolarSatelliteID, name2id("Satélite solar"))
+	assert.Equal(t, ID(0), name2id("人中位"))
+
+	// cz
+	assert.Equal(t, LightFighterID, name2id("Lehký stíhač"))
+	assert.Equal(t, HeavyFighterID, name2id("Těžký stíhač"))
+	assert.Equal(t, CruiserID, name2id("Křižník"))
+	assert.Equal(t, BattleshipID, name2id("Bitevní loď"))
+	assert.Equal(t, BattlecruiserID, name2id("Bitevní křižník"))
+	assert.Equal(t, BomberID, name2id("Bombardér"))
+	assert.Equal(t, DestroyerID, name2id("Ničitel"))
+	assert.Equal(t, DeathstarID, name2id("Hvězda smrti"))
+	assert.Equal(t, SmallCargoID, name2id("Malý transportér"))
+	assert.Equal(t, LargeCargoID, name2id("Velký transportér"))
+	assert.Equal(t, ColonyShipID, name2id("Kolonizační loď"))
+	assert.Equal(t, RecyclerID, name2id("Recyklátor"))
+	assert.Equal(t, EspionageProbeID, name2id("Špionážní sonda"))
+	assert.Equal(t, SolarSatelliteID, name2id("Solární satelit"))
+
+	// mx
+	assert.Equal(t, LightFighterID, name2id("Cazador ligero"))
+	assert.Equal(t, HeavyFighterID, name2id("Cazador pesado"))
+	assert.Equal(t, CruiserID, name2id("Crucero"))
+	assert.Equal(t, BattleshipID, name2id("Nave de batalla"))
+	assert.Equal(t, BattlecruiserID, name2id("Acorazado"))
+	assert.Equal(t, BomberID, name2id("Bombardero"))
+	assert.Equal(t, DestroyerID, name2id("Destructor"))
+	assert.Equal(t, DeathstarID, name2id("Estrella de la muerte"))
+	assert.Equal(t, SmallCargoID, name2id("Nave pequeña de carga"))
+	assert.Equal(t, LargeCargoID, name2id("Nave grande de carga"))
+	assert.Equal(t, ColonyShipID, name2id("Nave de la colonia"))
+	assert.Equal(t, RecyclerID, name2id("Reciclador"))
+	assert.Equal(t, EspionageProbeID, name2id("Sonda de espionaje"))
+	assert.Equal(t, SolarSatelliteID, name2id("Satélite solar"))
+
+	// hr
+	assert.Equal(t, LightFighterID, name2id("Mali lovac"))
+	assert.Equal(t, HeavyFighterID, name2id("Veliki lovac"))
+	assert.Equal(t, CruiserID, name2id("Krstarica"))
+	assert.Equal(t, BattleshipID, name2id("Borbeni brod"))
+	assert.Equal(t, BattlecruiserID, name2id("Oklopna krstarica"))
+	assert.Equal(t, BomberID, name2id("Bombarder"))
+	assert.Equal(t, DestroyerID, name2id("Razarač"))
+	assert.Equal(t, DeathstarID, name2id("Zvijezda smrti"))
+	assert.Equal(t, SmallCargoID, name2id("Mali transporter"))
+	assert.Equal(t, LargeCargoID, name2id("Veliki transporter"))
+	assert.Equal(t, ColonyShipID, name2id("Kolonijalni brod"))
+	assert.Equal(t, RecyclerID, name2id("Recikler"))
+	assert.Equal(t, EspionageProbeID, name2id("Sonde za špijunažu"))
+	assert.Equal(t, SolarSatelliteID, name2id("Solarni satelit"))
+
+	// it
+	assert.Equal(t, LightFighterID, name2id("Caccia Leggero"))
+	assert.Equal(t, HeavyFighterID, name2id("Caccia Pesante"))
+	assert.Equal(t, CruiserID, name2id("Incrociatore"))
+	assert.Equal(t, BattleshipID, name2id("Nave da battaglia"))
+	assert.Equal(t, BattlecruiserID, name2id("Incrociatore da Battaglia"))
+	assert.Equal(t, BomberID, name2id("Bombardiere"))
+	assert.Equal(t, DestroyerID, name2id("Corazzata"))
+	assert.Equal(t, DeathstarID, name2id("Morte Nera"))
+	assert.Equal(t, SmallCargoID, name2id("Cargo leggero"))
+	assert.Equal(t, LargeCargoID, name2id("Cargo Pesante"))
+	assert.Equal(t, ColonyShipID, name2id("Colonizzatrice"))
+	assert.Equal(t, RecyclerID, name2id("Riciclatrici"))
+	assert.Equal(t, EspionageProbeID, name2id("Sonda spia"))
+	assert.Equal(t, SolarSatelliteID, name2id("Satellite Solare"))
+
+	// pl
+	assert.Equal(t, LightFighterID, name2id("Lekki myśliwiec"))
+	assert.Equal(t, HeavyFighterID, name2id("Ciężki myśliwiec"))
+	assert.Equal(t, CruiserID, name2id("Krążownik"))
+	assert.Equal(t, BattleshipID, name2id("Okręt wojenny"))
+	assert.Equal(t, BattlecruiserID, name2id("Pancernik"))
+	assert.Equal(t, BomberID, name2id("Bombowiec"))
+	assert.Equal(t, DestroyerID, name2id("Niszczyciel"))
+	assert.Equal(t, DeathstarID, name2id("Gwiazda Śmierci"))
+	assert.Equal(t, SmallCargoID, name2id("Mały transporter"))
+	assert.Equal(t, LargeCargoID, name2id("Duży transporter"))
+	assert.Equal(t, ColonyShipID, name2id("Statek kolonizacyjny"))
+	assert.Equal(t, RecyclerID, name2id("Recykler"))
+	assert.Equal(t, EspionageProbeID, name2id("Sonda szpiegowska"))
+	assert.Equal(t, SolarSatelliteID, name2id("Satelita słoneczny"))
+
+	// tr
+	assert.Equal(t, LightFighterID, name2id("Hafif Avcı"))
+	assert.Equal(t, HeavyFighterID, name2id("Ağır Avcı"))
+	assert.Equal(t, CruiserID, name2id("Kruvazör Adet"))
+	assert.Equal(t, BattleshipID, name2id("Komuta Gemisi"))
+	assert.Equal(t, BattlecruiserID, name2id("Firkateyn"))
+	assert.Equal(t, BomberID, name2id("Bombardıman Gemisi"))
+	assert.Equal(t, DestroyerID, name2id("Muhrip"))
+	assert.Equal(t, DeathstarID, name2id("Ölüm Yildizi"))
+	assert.Equal(t, SmallCargoID, name2id("Küçük Nakliye Gemisi"))
+	assert.Equal(t, LargeCargoID, name2id("Büyük Nakliye Gemisi"))
+	assert.Equal(t, ColonyShipID, name2id("Koloni Gemisi"))
+	assert.Equal(t, RecyclerID, name2id("Geri Dönüsümcü"))
+	assert.Equal(t, EspionageProbeID, name2id("Casus Sondasi"))
+	assert.Equal(t, SolarSatelliteID, name2id("Solar Uydu"))
+
+	// ar
+	assert.Equal(t, LightFighterID, name2id("Cazador ligero"))
+	assert.Equal(t, HeavyFighterID, name2id("Cazador pesado"))
+	assert.Equal(t, CruiserID, name2id("Crucero"))
+	assert.Equal(t, BattleshipID, name2id("Nave de batalla"))
+	assert.Equal(t, BattlecruiserID, name2id("Acorazado"))
+	assert.Equal(t, BomberID, name2id("Bombardero"))
+	assert.Equal(t, DestroyerID, name2id("Destructor"))
+	assert.Equal(t, DeathstarID, name2id("Estrella de la muerte"))
+	assert.Equal(t, SmallCargoID, name2id("Nave pequeña de carga"))
+	assert.Equal(t, LargeCargoID, name2id("Nave grande de carga"))
+	assert.Equal(t, ColonyShipID, name2id("Nave colonizadora"))
+	assert.Equal(t, RecyclerID, name2id("Reciclador"))
+	assert.Equal(t, EspionageProbeID, name2id("Sonda de espionaje"))
+	assert.Equal(t, SolarSatelliteID, name2id("Satélite solar"))
+
+	//pt
+	assert.Equal(t, LightFighterID, name2id("Caça Ligeiro"))
+	assert.Equal(t, HeavyFighterID, name2id("Caça Pesado"))
+	assert.Equal(t, CruiserID, name2id("Cruzador"))
+	assert.Equal(t, BattleshipID, name2id("Nave de Batalha"))
+	assert.Equal(t, BattlecruiserID, name2id("Interceptor"))
+	assert.Equal(t, BomberID, name2id("Bombardeiro"))
+	assert.Equal(t, DestroyerID, name2id("Destruidor"))
+	assert.Equal(t, DeathstarID, name2id("Estrela da Morte"))
+	assert.Equal(t, SmallCargoID, name2id("Cargueiro Pequeno"))
+	assert.Equal(t, LargeCargoID, name2id("Cargueiro Grande"))
+	assert.Equal(t, ColonyShipID, name2id("Nave de Colonização"))
+	assert.Equal(t, RecyclerID, name2id("Reciclador"))
+	assert.Equal(t, EspionageProbeID, name2id("Sonda de Espionagem"))
+	assert.Equal(t, SolarSatelliteID, name2id("Satélite Solar"))
+
+	//nl
+	assert.Equal(t, LightFighterID, name2id("Licht gevechtsschip"))
+	assert.Equal(t, HeavyFighterID, name2id("Zwaar gevechtsschip"))
+	assert.Equal(t, CruiserID, name2id("Kruiser"))
+	assert.Equal(t, BattleshipID, name2id("Slagschip"))
+	assert.Equal(t, BattlecruiserID, name2id("Interceptor"))
+	assert.Equal(t, BomberID, name2id("Bommenwerper"))
+	assert.Equal(t, DestroyerID, name2id("Vernietiger"))
+	assert.Equal(t, DeathstarID, name2id("Ster des Doods"))
+	assert.Equal(t, SmallCargoID, name2id("Klein vrachtschip"))
+	assert.Equal(t, LargeCargoID, name2id("Groot vrachtschip"))
+	assert.Equal(t, ColonyShipID, name2id("Kolonisatieschip"))
+	assert.Equal(t, RecyclerID, name2id("Recycler"))
+	assert.Equal(t, EspionageProbeID, name2id("Spionagesonde"))
+	assert.Equal(t, SolarSatelliteID, name2id("Zonne-energiesatelliet"))
+
+	//dk
+	assert.Equal(t, LightFighterID, name2id("Lille Jæger"))
+	assert.Equal(t, HeavyFighterID, name2id("Stor Jæger"))
+	assert.Equal(t, CruiserID, name2id("Krydser"))
+	assert.Equal(t, BattleshipID, name2id("Slagskib"))
+	assert.Equal(t, BattlecruiserID, name2id("Interceptor"))
+	assert.Equal(t, BomberID, name2id("Bomber"))
+	assert.Equal(t, DestroyerID, name2id("Destroyer"))
+	assert.Equal(t, DeathstarID, name2id("Dødsstjerne"))
+	assert.Equal(t, SmallCargoID, name2id("Lille Transporter"))
+	assert.Equal(t, LargeCargoID, name2id("Stor Transporter"))
+	assert.Equal(t, ColonyShipID, name2id("Koloniskib"))
+	assert.Equal(t, RecyclerID, name2id("Recycler"))
+	assert.Equal(t, EspionageProbeID, name2id("Spionagesonde"))
+	assert.Equal(t, SolarSatelliteID, name2id("Solarsatellit"))
+
+	// ru
+	assert.Equal(t, LightFighterID, name2id("Лёгкий истребитель"))
+	assert.Equal(t, HeavyFighterID, name2id("Тяжёлый истребитель"))
+	assert.Equal(t, CruiserID, name2id("Крейсер"))
+	assert.Equal(t, BattleshipID, name2id("Линкор"))
+	assert.Equal(t, BattlecruiserID, name2id("Линейный крейсер"))
+	assert.Equal(t, BomberID, name2id("Бомбардировщик"))
+	assert.Equal(t, DestroyerID, name2id("Уничтожитель"))
+	assert.Equal(t, DeathstarID, name2id("Звезда смерти"))
+	assert.Equal(t, SmallCargoID, name2id("Малый транспорт"))
+	assert.Equal(t, LargeCargoID, name2id("Большой транспорт"))
+	assert.Equal(t, ColonyShipID, name2id("Колонизатор"))
+	assert.Equal(t, RecyclerID, name2id("Переработчик"))
+	assert.Equal(t, EspionageProbeID, name2id("Шпионский зонд"))
+	assert.Equal(t, SolarSatelliteID, name2id("Солнечный спутник"))
+
+	// jp
+	assert.Equal(t, LightFighterID, name2id("軽戦闘機"))
+	assert.Equal(t, HeavyFighterID, name2id("重戦闘機"))
+	assert.Equal(t, CruiserID, name2id("巡洋艦"))
+	assert.Equal(t, BattleshipID, name2id("バトルシップ"))
+	assert.Equal(t, BattlecruiserID, name2id("大型戦艦"))
+	assert.Equal(t, BomberID, name2id("爆撃機"))
+	assert.Equal(t, DestroyerID, name2id("デストロイヤー"))
+	assert.Equal(t, DeathstarID, name2id("デススター"))
+	assert.Equal(t, SmallCargoID, name2id("小型輸送機"))
+	assert.Equal(t, LargeCargoID, name2id("大型輸送機"))
+	assert.Equal(t, ColonyShipID, name2id("コロニーシップ"))
+	assert.Equal(t, RecyclerID, name2id("残骸回収船"))
+	assert.Equal(t, EspionageProbeID, name2id("偵察機"))
+	assert.Equal(t, SolarSatelliteID, name2id("ソーラーサテライト"))
+}
